requirements.txt
Django == 2.0.0
djangorestframework == 3.8.2
django-filter == 1.1.0